+++ backend/package.json
@@ -6,6 +6,8 @@
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
 		"dev": "nodemon src/app.ts",
+		"build": "npx tsc",
+		"start": "node ./dist/app.js",
 		"db:migrate": "npx prisma migrate dev --name latest",
 		"db:reset": "npx prisma migrate reset"
 	},
