+++ backend/src/controllers/Project.controller.ts
@@ -1,26 +1,27 @@
 import { Request, Response } from "express";
-import { PrismaClient } from "@prisma/client";
-import ResponseSender from "../utils/responseSender";
-import { projectMessages, systemMessages, userMessages } from "../utils/messages";
-import InvalidChecker from "../utils/invalidChecker";
 
+import ResponseUtils, { Messages } from "../utils/Response.utils";
+import SystemUtils from "../utils/System.utils";
+import ProjectUtils from "../utils/Project.utils";
+
+import { PrismaClient } from "@prisma/client";
 const prisma = new PrismaClient();
 
-const { error: projErr, success: projSuc } = projectMessages;
-const { error: userErr } = userMessages;
-const { error: sysErr } = systemMessages;
+const { error: projErr, success: projSuc } = Messages.projectMessages;
+const { error: userErr } = Messages.userMessages;
+const { error: sysErr } = Messages.systemMessages;
 
-export default class ProductController {
+export default class ProjectController {
 	static async getUserProjects(req: Request, res: Response) {
 		// Checa se o usuário está logado
 		if (!req.session.user) {
-			return ResponseSender.sendMessage(userErr.notLoggedYet, req, res);
+			return ResponseUtils.sendMessage(userErr.notLoggedYet, req, res);
 		}
 
-		const projects = await prisma.project.findMany({ where: { userId: req.session.user.id }, select: { id: true, name: true, desc: true } });
 		// Checa se o usuário tem projetos
+		const projects = await prisma.project.findMany({ where: { userId: req.session.user.id }, select: { id: true, name: true, desc: true } });
 		if (!projects || !projects.length) {
-			return ResponseSender.sendMessage(projErr.projectsNotFound, req, res);
+			return ResponseUtils.sendMessage(projErr.projectsNotFound, req, res);
 		}
 
 		// Retorna os projetos do usuário
@@ -30,13 +31,13 @@ export default class ProductController {
 	static async createProject(req: Request, res: Response) {
 		// Checa se o usuário tem permissão de criar um projeto, ou seja, se está logado
 		if (!req.session.user) {
-			return ResponseSender.sendMessage(userErr.notLoggedYet, req, res);
+			return ResponseUtils.sendMessage(userErr.notLoggedYet, req, res);
 		}
 
 		// Checa se os dados enviados estão vazios
 		const { name, desc } = req.body;
-		if (InvalidChecker.isEmpty(name, desc)) {
-			return ResponseSender.sendMessage(sysErr.emptyValues, req, res);
+		if (SystemUtils.isEmpty(name, desc)) {
+			return ResponseUtils.sendMessage(sysErr.emptyValues, req, res);
 		}
 
 		// Cria um projeto para o usuário
@@ -52,25 +53,25 @@ export default class ProductController {
 			},
 		});
 
-		ResponseSender.sendMessage(projSuc.projectCreated, req, res);
+		ResponseUtils.sendMessage(projSuc.projectCreated, req, res);
 	}
 
 	static async updateName(req: Request, res: Response) {
 		// Checa se o usuário está logado
 		if (!req.session.user) {
-			return ResponseSender.sendMessage(userErr.notLoggedYet, req, res);
+			return ResponseUtils.sendMessage(userErr.notLoggedYet, req, res);
 		}
 
 		// Checa se o nome e ID está vazio
 		const { name, id } = req.body;
-		if (InvalidChecker.isEmpty(name, id)) {
-			return ResponseSender.sendMessage(sysErr.emptyValues, req, res);
+		if (SystemUtils.isEmpty(name, id)) {
+			return ResponseUtils.sendMessage(sysErr.emptyValues, req, res);
 		}
 
 		// Verifica se o projeto é do usuário logado, para evitar erros de inexistência e edição de projetos alheios
-		const project = await prisma.project.findFirst({ where: { AND: [{ userId: req.session.user.id }, { id }] } });
-		if (!project) {
-			return ResponseSender.sendMessage(projErr.notOwner, req, res);
+		const isOwner = await ProjectUtils.isUserOwner(req.session.user, id, prisma);
+		if (!isOwner) {
+			return ResponseUtils.sendMessage(projErr.notOwner, req, res);
 		}
 
 		// Atualiza o nome do projeto
@@ -86,25 +87,25 @@ export default class ProductController {
 			},
 		});
 
-		ResponseSender.sendMessage(projSuc.projectNameUpdated, req, res);
+		ResponseUtils.sendMessage(projSuc.projectNameUpdated, req, res);
 	}
 
 	static async updateDescription(req: Request, res: Response) {
 		// Checa se o usuário está logado
 		if (!req.session.user) {
-			return ResponseSender.sendMessage(userErr.notLoggedYet, req, res);
+			return ResponseUtils.sendMessage(userErr.notLoggedYet, req, res);
 		}
 
 		// Checa se a descrição e ID está vazio
 		const { desc, id } = req.body;
-		if (InvalidChecker.isEmpty(desc, id)) {
-			return ResponseSender.sendMessage(sysErr.emptyValues, req, res);
+		if (SystemUtils.isEmpty(desc, id)) {
+			return ResponseUtils.sendMessage(sysErr.emptyValues, req, res);
 		}
 
 		// Verifica se o projeto é do usuário logado, para evitar erros de inexistência e edição de projetos alheios
-		const project = await prisma.project.findFirst({ where: { AND: [{ userId: req.session.user.id }, { id }] } });
-		if (!project) {
-			return ResponseSender.sendMessage(projErr.notOwner, req, res);
+		const isOwner = await ProjectUtils.isUserOwner(req.session.user, id, prisma);
+		if (!isOwner) {
+			return ResponseUtils.sendMessage(projErr.notOwner, req, res);
 		}
 
 		// Atualiza o nome do projeto
@@ -120,25 +121,25 @@ export default class ProductController {
 			},
 		});
 
-		ResponseSender.sendMessage(projSuc.projectDescUpdated, req, res);
+		ResponseUtils.sendMessage(projSuc.projectDescUpdated, req, res);
 	}
 
 	static async deleteProject(req: Request, res: Response) {
 		// Checa se o usuário está logado
 		if (!req.session.user) {
-			return ResponseSender.sendMessage(userErr.notLoggedYet, req, res);
+			return ResponseUtils.sendMessage(userErr.notLoggedYet, req, res);
 		}
 
 		// Checa se o ID do projeto está vazio
 		const { id } = req.body;
 		if (!id) {
-			return ResponseSender.sendMessage(sysErr.emptyValues, req, res);
+			return ResponseUtils.sendMessage(sysErr.emptyValues, req, res);
 		}
 
 		// Verifica se o projeto é do usuário logado, para evitar erros de inexistência e deleção de projetos alheios
-		const project = await prisma.project.findFirst({ where: { AND: [{ userId: req.session.user.id }, { id }] } });
-		if (!project) {
-			return ResponseSender.sendMessage(projErr.notOwner, req, res);
+		const isOwner = await ProjectUtils.isUserOwner(req.session.user, id, prisma);
+		if (!isOwner) {
+			return ResponseUtils.sendMessage(projErr.notOwner, req, res);
 		}
 
 		await prisma.user.update({
@@ -152,6 +153,6 @@ export default class ProductController {
 			},
 		});
 
-		ResponseSender.sendMessage(projSuc.projectDeleted, req, res);
+		ResponseUtils.sendMessage(projSuc.projectDeleted, req, res);
 	}
 }
