+++ backend/src/routes/User.routes.ts
@@ -1,5 +1,6 @@
 import { Router } from "express";
 import UserController from "../controllers/User.controller";
+
 const router = Router();
 
 router.post("/register", UserController.registerUser);
