+++ backend/src/types/Message.type.ts
@@ -1,0 +1,8 @@
+export type Message = {
+	status: number;
+	json: {
+		error?: string;
+		success?: string;
+		code: string;
+	};
+};
