+++ backend/src/types/Session.type.ts
@@ -2,6 +2,12 @@ import "express-session";
 
 declare module "express-session" {
 	export interface SessionData {
-		user: { id: number; email: string; username: string };
+		user: User;
 	}
 }
+
+export type User = {
+	id: number;
+	email: string;
+	username: string;
+};
