+++ backend/src/utils/System.utils.ts
@@ -1,6 +1,6 @@
-export default class InvalidChecker {
+export default class SystemUtils {
 	// Checa se algum dos argumentos passados estão vazios, retornando true caso algum esteja
-	static isEmpty(...args: String[]): boolean {
+	static isEmpty(...args: any[]): boolean {
 		const arr: boolean[] = args.map((item) => {
 			return !item ? true : false;
 		});
