+++ backend/src/utils/User.utils.ts
@@ -1,4 +1,3 @@
-import { Request } from "express";
 import bcrypt from "bcrypt";
 
 export default class UserUtils {
