+++ backend/tsconfig.json
@@ -6,6 +6,8 @@
 		"forceConsistentCasingInFileNames": true,
 		"strict": true,
 		"skipLibCheck": true,
+		"rootDir": "./src",
+		"outDir": "./dist",
 		"typeRoots": ["./node_modules/@types", "./src/types"]
 	}
 }
