--- frontend/.env
@@ -1,2 +1,0 @@
-BASE_URL=/
-VITE_API_REST_URL=http://localhost:3000/
