+++ frontend/package.json
@@ -1,33 +1,26 @@
 {
-  "name": "conciliacao_bancaria",
+  "name": "frontend",
+  "version": "0.0.0",
   "private": true,
-  "version": "0.0.1",
-  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vue-tsc --noEmit && vite build",
-    "preview": "vite preview"
+    "build": "run-p type-check build-only",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --noEmit"
   },
   "dependencies": {
-    "@quasar/extras": "^1.14.3",
-    "apexcharts": "^3.35.5",
-    "axios": "^0.27.2",
-    "pinia": "^2.0.16",
-    "quasar": "^2.7.5",
-    "sortablejs": "1.10.2",
-    "vue": "^3.2.37",
-    "vue-draggable-next": "^2.1.1",
-    "vue-router": "^4.1.2",
-    "vue-sortable": "^0.1.3",
-    "vue-the-mask": "^0.11.1",
-    "vue3-smooth-dnd": "^0.0.2",
-    "vuedraggable": "^2.24.3"
+    "axios": "^1.4.0",
+    "vue": "^3.2.47",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
-    "@quasar/vite-plugin": "^1.0.10",
-    "@vitejs/plugin-vue": "^3.0.0",
-    "typescript": "^4.6.4",
-    "vite": "^3.0.0",
-    "vue-tsc": "^0.38.4"
+    "@types/node": "^18.14.2",
+    "@vitejs/plugin-vue": "^4.0.0",
+    "@vue/tsconfig": "^0.1.3",
+    "npm-run-all": "^4.1.5",
+    "typescript": "~4.8.4",
+    "vite": "^4.1.4",
+    "vue-tsc": "^1.2.0"
   }
 }
