+++ frontend/src/App.vue
@@ -1,15 +1,8 @@
+<script setup lang="ts"></script>
+
 <template>
-  <div>
-   <!-- <MenuTemplate /> -->
-   <router-view></router-view>
-  </div>
-  
+	<div>
+		<h1>Frontend</h1>
+		<RouterView />
+	</div>
 </template>
-
-<script setup lang="ts">
-// import MenuTemplate from './components/MenuTemplate.vue'
-
-</script>
-
-
-
