--- frontend/src/core/types/IKResult.ts
@@ -1,11 +1,0 @@
-export interface IKResult<T> {
-    succeeded: boolean,
-    eRRORs: IKResultError[],
-    result: T
-}
-
-
-export interface IKResultError {
-    code: String,
-    description: String
-}
