--- frontend/src/css/config.css
@@ -1,19 +1,0 @@
-.Black{
- background-color: #231f20;
-}
-
-.BlackGray{
-  background: #353535;
-}
-
-.Purple{
-  background-color: #5E477A;;
-}
-
-.BgInput{
-  background-color: #959595;
-}
-
-.btn {
-  width: 100%;
-}
