+++ frontend/vite.config.ts
@@ -1,28 +1,14 @@
+import { fileURLToPath, URL } from 'node:url'
+
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
-import path from 'path'
-import { quasar, transformAssetUrls } from '@quasar/vite-plugin'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [
-    vue({
-      template: { transformAssetUrls }
-    }),
-
-    quasar()
-  ],
-
-
+  plugins: [vue()],
   resolve: {
     alias: {
-      '@': path.resolve(__dirname, './src'),
-      '@components': path.resolve(__dirname, './src/components'),
-      '@pages': path.resolve(__dirname, './src/pages'),
-      '@modules': path.resolve(__dirname, './src/modules'),
-      
-      '@store': path.resolve(__dirname, './src/core/store'),
-      '@types': path.resolve(__dirname, './src/core/types'),
+      '@': fileURLToPath(new URL('./src', import.meta.url))
     }
   }
 })
